% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_pois.R
\name{experiment_pois}
\alias{experiment_pois}
\title{Simulador de experimentos}
\usage{
experiment_pois(configuracoes)
}
\arguments{
\item{configuracoes}{Uma lista de configurações (listas nomeadas) para os experimentos, com valores}
}
\value{
Um `data frame` com as colunas:
\itemize{
\item exper: O índice do experimento.
\item i: O índice da réplica (dentro do experimento).
\item emv: O valor do EMV para a simulação do índice i.
\item n: O valor de n utilizado no experimento.
\item lambda: O valor de λ utilizado no experimento.
}
}
\description{
Simulador de experimentos
}
