% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafs_pois.R
\name{graf_pois}
\alias{graf_pois}
\title{Gráficos de densidade do EMV da distribuição de Poisson.}
\usage{
graf_pois(input)
}
\arguments{
\item{input:}{Resultado obtidos pela função `experiment_pois(lista)`}
}
\value{
Gráfico da distribuição do EMV (empírica obtida através de
simulações) para diferentes valores de \eqn{\lambda} e tamanhos de
amostra `n`.
}
\description{
Gráficos de densidade do EMV da distribuição de Poisson.
}
\examples{
configs <- list(
  list(n = 10, n_replicas = 200, lambda = 5),
  list(n = 50, n_replicas = 200, lambda = 5),
  list(n = 100, n_replicas = 200, lambda = 5),
  list(n = 10, n_replicas = 200, lambda = 20),
  list(n = 50, n_replicas = 200, lambda = 20),
  list(n = 100, n_replicas = 200, lambda = 20))
input <- experiment_pois(configs)
graf_pois(input)
}
