% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rpois.R
\name{list_rpois}
\alias{list_rpois}
\title{Lista de Amostras Poisson}
\usage{
list_rpois(n, n_replicas, lambda)
}
\arguments{
\item{n}{Tamanho das amostras}

\item{n_replicas}{Número de Amostras}

\item{lambda}{Valor de \eqn{\lambda} utilizado nas amostras.}
}
\value{
Uma lista com `n_replicas` amostras de tamanho `n` da poisson \eqn{\lambda}.
}
\description{
Lista de Amostras Poisson
}
\examples{
list_rpois(n = 10, n_replicas = 2, lambda = 4)
list_rpois(n = 5, n_replicas = 5, lambda = 5)

}
