% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_extras2.R
\name{pois_extras2}
\alias{pois_extras2}
\title{Combinações de `n` amostra, \eqn{\lambda} e `replicas`.}
\usage{
pois_extras2(n_p, lambda_p, n_replicas_p)
}
\arguments{
\item{lambda_p}{vetor com valores de \eqn{\lambda}.}

\item{n_replicas_p}{Único valor de replicas.}

\item{n_p:}{vetor com valores de `n` amostra.}
}
\value{
Lista com todas as combinações de `n` amostra,
\eqn{\lambda} e `replicas`.
}
\description{
Combinações de `n` amostra, \eqn{\lambda} e `replicas`.
}
\examples{
pois_extras2(n_p=c(10, 50, 100),
             lambda_p=c(5, 20),
             n_replicas_p=2)
}
