% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preguica_de_listar.R
\name{preguica_de_listar}
\alias{preguica_de_listar}
\title{Cria todas as listas com todas as combinações `n`s e \eqn{\lambda}s informadas}
\usage{
preguica_de_listar(n, lambda, n_replicas)
}
\arguments{
\item{n}{Vetor numérico com os tamanhos de amostra}

\item{lambda}{Vetor numérico com os \eqn{\lambda}s da Poisson}

\item{n_replicas}{Número de réplicas em cada experimento}
}
\description{
Cria todas as listas com todas as combinações `n`s e \eqn{\lambda}s informadas
}
