% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiza_emv.R
\name{visualiza_emv}
\alias{visualiza_emv}
\title{Visualizador dos EMV de \eqn{\lambda} para Poisson}
\usage{
visualiza_emv(data)
}
\arguments{
\item{data}{um `data frame` no formato de saída da função experiment_pois}
}
\value{
Um gráfico feito no ggplot2 com a visualização dos dados
}
\description{
Visualizador dos EMV de \eqn{\lambda} para Poisson
}
