% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_pois.R
\name{experiment_pois}
\alias{experiment_pois}
\title{Resultados de amostras do EMV.}
\usage{
experiment_pois(listas)
}
\arguments{
\item{list:}{Lista ou um conjunto de lista com
amostra desejavél para simular.}

\item{n:}{Tamanho das amostras}

\item{n_replicas:}{Número de réplicas da simulação.}

\item{lambda:}{Parâmetro utilizado na simulação.}
}
\value{
resultados de amostras do Estimador de Máxima Verossimilhança
para um par de `n` e \eqn{\lambda}.
}
\description{
Resultados de amostras do EMV.
}
\examples{
configs <- list(
  list(n = 10, n_replicas = 2, lambda = 3.14),
  list(n = 50, n_replicas = 3, lambda = 3.14),
  list(n = 50, n_replicas = 2, lambda = 20))
experiment_pois(configs)
}
