% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rpois.R
\name{list_rpois}
\alias{list_rpois}
\title{Gera uma lista com amostras da distribuição Poisson}
\usage{
list_rpois(n, n_replicas, lambda)
}
\arguments{
\item{n}{tamanho das amostras;}

\item{n_replicas}{número de réplicas da simulação;}

\item{lambda}{Parâmetro utilizado na simulação.}
}
\value{
A função deve retorna uma lista com n_replicas elementos, sendo cada elemento uma amostra de tamanho n (vetor) da distribuição Poisson com parâmetro lambda.
}
\description{
Gera uma lista com amostras da distribuição Poisson
}
