% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rpois.R
\name{list_rpois}
\alias{list_rpois}
\title{Amostra da distribuição Poisson}
\usage{
list_rpois(n, n_replicas, lambda)
}
\arguments{
\item{n:}{Tamanho das amostras}

\item{n_replicas:}{Número de réplicas da simulação.}

\item{lambda:}{Parâmetro utilizado na simulação.}
}
\value{
listas de replicações da distribuição de Poisson.
}
\description{
A função deve retornar uma lista com `n_replicas` elementos, sendo cada
elemento uma amostra de tamanho `n`(vetor) da distribuição Poisson
com parâmetro lambda.
}
\examples{
list_rpois(n = 10, n_replicas = 2, lambda = 4)

}
